Update nsd controller and routes
