--- meao/utils/startup.py
@@ -1,25 +1,0 @@
-# copy paste from main branch, solve on merging
-
-import json
-
-import cherrypy
-
-
-def jsonify_error(status, message, traceback, version):
-    """JSONify all CherryPy error responses (created by raising the
-    cherrypy.HTTPError exception)
-    """
-
-    cherrypy.response.headers["Content-Type"] = "application/json"
-    response_body = json.dumps(
-        {
-            "error": {
-                "http_status": status,
-                "message": message,
-            }
-        }
-    )
-
-    cherrypy.response.status = status
-
-    return response_body
